Simple Premade Tripwire
## About The Pull Request
Tripwire traps are a structure ive just been considering for a while.
Tripwire traps can be mapload or crafted.
If generated at mapload any /obj/item in their loc will be used as the attack.
If crafted they will be empty until a item is used on them.
When Crossed() the item will be forcemoved to the turf and throw_impact will be called on tthe crosser.
If a grenade is loaded into the tripwire trap it will arm the grenade.
When a tripwire trap breaks it produces 1 cloth. This is to ensure a net loss in material if someone wants to mass produce them.
Functionally they would be similar to Dwarf Fortress rock traps that drop a single rock item onto someone.
## Why It's Good For The Game
It would be helpful for making custom traps in city maps instead of relying on contact mines.
## Changelog
🆑
add: Tripwire Traps
/🆑